Fixed hashCode float to int implicit conversion.

+++ source/include/G3D/Sphere.h
@@ -120,7 +120,7 @@ public:
 } // namespace
 
 inline unsigned int hashCode(const G3D::Sphere& sphere) {
-    return (hashCode(sphere.center) + (sphere.radius * 13));
+    return (unsigned int)(hashCode(sphere.center) + (sphere.radius * 13));
 }
 
 #endif
